+++ packages/react/src/atomic/Link/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Link component allows us to navigate to external addresses.
 
+## 2023-02-24 `3.5.1`
+
+### 💡 Others
+
+- Adjust `fontSize` and `color` properties in `Text` component on stories documentation. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2023-02-23 `3.5.0`
 
 ### 🎉 New features