+++ packages/react/src/atomic/Text/CHANGELOG.md
@@ -7,6 +7,7 @@ Text is a basic component that allows us to write blocks of text and give it for
 ### 🎉 New features
 
 - Changed `fontSize` default value to `base`. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+- Changed `lineHeight` default value to `base`. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
 - Changed `color` default value to `neutral.textLow`. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
 
 ## 2023-02-23 `5.1.0`